Dedupe more code
