+++ plugin-server/src/cdp/hog-transformations/hog-transformer.service.ts
@@ -7,17 +7,17 @@ import {
     HogFunctionType,
     HogFunctionTypeType,
 } from '../../cdp/types'
-import { checkHogFunctionFilters, createInvocation, isLegacyPluginHogFunction } from '../../cdp/utils'
+import { createInvocation, isLegacyPluginHogFunction } from '../../cdp/utils'
 import { runInstrumentedFunction } from '../../main/utils'
 import { Hub } from '../../types'
 import { logger } from '../../utils/logger'
 import { buildGlobalsWithInputs, HogExecutorService } from '../services/hog-executor.service'
-import { hogFunctionFilterDuration } from '../services/hog-executor.service'
 import { HogFunctionManagerService } from '../services/hog-function-manager.service'
 import { HogFunctionManagerLazyService } from '../services/hog-function-manager-lazy.service'
 import { HogFunctionMonitoringService } from '../services/hog-function-monitoring.service'
 import { LegacyPluginExecutorService } from '../services/legacy-plugin-executor.service'
 import { convertToHogFunctionFilterGlobal } from '../utils'
+import { checkHogFunctionFilters } from '../utils/hog-function-filtering'
 import { cleanNullValues } from './transformation-functions'
 
 export const hogTransformationDroppedEvents = new Counter({
@@ -168,9 +168,6 @@ export class HogTransformerService {
                                     filterGlobals,
                                     eventUuid: globals.event?.uuid,
                                 })
-
-                                // Track the duration for metrics
-                                hogFunctionFilterDuration.observe({ type: hogFunction.type }, filterResults.duration)
 
                                 // If filter didn't pass and there was no error, skip this transformation
                                 if (!filterResults.match && !filterResults.error) {
