+++ plugin-server/src/cdp/services/hog-executor.service.ts
@@ -22,12 +22,8 @@ import {
     HogFunctionQueueParametersFetchResponse,
     HogFunctionType,
 } from '../types'
-import {
-    buildExportedFunctionInvoker,
-    checkHogFunctionFilters,
-    convertToHogFunctionFilterGlobal,
-    createInvocation,
-} from '../utils'
+import { buildExportedFunctionInvoker, convertToHogFunctionFilterGlobal, createInvocation } from '../utils'
+import { checkHogFunctionFilters } from '../utils/hog-function-filtering'
 
 export const MAX_ASYNC_STEPS = 5
 export const MAX_HOG_LOGS = 25
@@ -39,14 +35,6 @@ const hogExecutionDuration = new Histogram({
     help: 'Processing time and success status of internal functions',
     // We have a timeout so we don't need to worry about much more than that
     buckets: [0, 10, 20, 50, 100, 200],
-})
-
-export const hogFunctionFilterDuration = new Histogram({
-    name: 'cdp_hog_function_filter_duration_ms',
-    help: 'Processing time for filtering a function',
-    // We have a timeout so we don't need to worry about much more than that
-    buckets: [0, 10, 20, 50, 100, 200],
-    labelNames: ['type'],
 })
 
 const hogFunctionStateMemory = new Histogram({
@@ -181,9 +169,6 @@ export class HogExecutorService {
             // Add any generated metrics and logs to our collections
             metrics.push(...filterResults.metrics)
             logs.push(...filterResults.logs)
-
-            // Record the duration for monitoring
-            hogFunctionFilterDuration.observe({ type: hogFunction.type }, filterResults.duration)
 
             return filterResults.match
         }
