+++ plugin-server/src/cdp/utils.ts
@@ -1,7 +1,7 @@
 // NOTE: PostIngestionEvent is our context event - it should never be sent directly to an output, but rather transformed into a lightweight schema
 
 import { CyclotronJob, CyclotronJobUpdate } from '@posthog/cyclotron'
-import { Bytecodes, ExecResult, HogVMException } from '@posthog/hogvm'
+import { Bytecodes } from '@posthog/hogvm'
 import { DateTime } from 'luxon'
 import RE2 from 're2'
 import { gunzip, gzip } from 'zlib'
@@ -14,9 +14,7 @@ import { captureException } from '../utils/posthog'
 import { castTimestampOrNow, clickHouseTimestampToISO, UUIDT } from '../utils/utils'
 import { MAX_GROUP_TYPES_PER_TEAM } from '../worker/ingestion/group-type-manager'
 import { CdpInternalEvent } from './schema'
-import { execHog } from './services/hog-executor.service'
 import {
-    HogFunctionAppMetric,
     HogFunctionCapturedEvent,
     HogFunctionFilterGlobals,
     HogFunctionInvocation,
@@ -24,7 +22,6 @@ import {
     HogFunctionInvocationGlobalsWithInputs,
     HogFunctionInvocationLogEntry,
     HogFunctionInvocationQueueParameters,
-    HogFunctionInvocationResult,
     HogFunctionInvocationSerialized,
     HogFunctionLogEntrySerialized,
     HogFunctionType,
@@ -462,117 +459,4 @@ export function buildExportedFunctionInvoker(
 
 export function isLegacyPluginHogFunction(hogFunction: HogFunctionType): boolean {
     return hogFunction.template_id?.startsWith('plugin-') ?? false
-}
-
-interface HogFunctionFilterResult {
-    match: boolean
-    error?: unknown
-    logs: HogFunctionInvocationLogEntry[]
-    metrics: HogFunctionAppMetric[]
-    duration: number
-}
-
-/**
- * Shared utility to check if an event matches the filters of a HogFunction.
- * Used by both the HogExecutorService (for destinations) and HogTransformerService (for transformations).
- */
-export function checkHogFunctionFilters(options: {
-    hogFunction: HogFunctionType
-    filterGlobals: HogFunctionFilterGlobals
-    /** Optional filters to use instead of those on the function */
-    filters?: HogFunctionType['filters']
-    /** Whether to enable telemetry for this function at the hogvm level */
-    enabledTelemetry?: boolean
-    /** The event UUID to use for logging */
-    eventUuid?: string
-}): HogFunctionFilterResult {
-    const { hogFunction, filterGlobals, enabledTelemetry, eventUuid } = options
-    const filters = options.filters ?? hogFunction.filters
-    const start = performance.now()
-    const logs: HogFunctionInvocationLogEntry[] = []
-    const metrics: HogFunctionAppMetric[] = []
-
-    let execResult: ExecResult | undefined
-    const result: HogFunctionFilterResult = {
-        match: false,
-        logs,
-        metrics,
-        duration: 0,
-    }
-
-    if (!filters?.bytecode) {
-        result.error = 'No filters bytecode'
-        return result
-    }
-
-    try {
-        execResult = execHog(filters.bytecode, {
-            globals: filterGlobals,
-            telemetry: enabledTelemetry,
-        })
-
-        if (execResult.error) {
-            throw execResult.error
-        }
-
-        result.match = typeof execResult.result === 'boolean' && execResult.result
-
-        if (!result.match) {
-            metrics.push({
-                team_id: hogFunction.team_id,
-                app_source_id: hogFunction.id,
-                metric_kind: 'other',
-                metric_name: 'filtered',
-                count: 1,
-            })
-        }
-    } catch (error) {
-        logger.error('🦔', `[HogFunction] Error filtering function`, {
-            hogFunctionId: hogFunction.id,
-            hogFunctionName: hogFunction.name,
-            teamId: hogFunction.team_id,
-            error: error.message,
-            result: execResult,
-        })
-
-        metrics.push({
-            team_id: hogFunction.team_id,
-            app_source_id: hogFunction.id,
-            metric_kind: 'other',
-            metric_name: 'filtering_failed',
-            count: 1,
-        })
-
-        if (eventUuid) {
-            logs.push({
-                team_id: hogFunction.team_id,
-                log_source: 'hog_function',
-                log_source_id: hogFunction.id,
-                instance_id: new UUIDT().toString(),
-                timestamp: DateTime.now(),
-                level: 'error',
-                message: `Error filtering event ${eventUuid}: ${error.message}`,
-            })
-        }
-        result.error = error.message
-    } finally {
-        const duration = performance.now() - start
-
-        // Re-using the constant from hog-executor.service.ts
-        const DEFAULT_TIMEOUT_MS = 100
-
-        if (duration > DEFAULT_TIMEOUT_MS) {
-            logger.error('🦔', `[HogFunction] Filter took longer than expected`, {
-                hogFunctionId: hogFunction.id,
-                hogFunctionName: hogFunction.name,
-                teamId: hogFunction.team_id,
-                duration,
-                eventId: options?.eventUuid,
-            })
-        }
-    }
-
-    result.duration = performance.now() - start
-
-    return result
 }
